FIXED: * Multiple row styles * Recipe only saves final direction step * Check rows are updated, not destroyed and recreated.

+++ database/migrations/2017_06_07_124636_create_rows_table.php
@@ -22,7 +22,7 @@ class CreateRowsTable extends Migration
             $table->integer('delta')->unsigned();
             $table->integer('unit_id')->unsigned();
             $table->foreign('unit_id')->references('id')->on('units');
-            $table->integer('value')->unsigned();
+            $table->float('value', 10, 5)->unsigned();
             $table->float('weight', 10, 5)->unsigned();
         });
     }
